refactor: update team dashboard components and styles
- Removed outline and box-shadow styles from global CSS for a cleaner look.
- Replaced TeamSearch with TeamDashboardHeader in the team page for improved navigation.
- Added TeamActivityCard component for displaying team activities.
- Updated TeamMembers component to use unique IDs for keys and improved button styling.
- Enhanced TeamTask component with loading and error states for better user experience.
- Updated TODO comments for clarity on future API integrations.
