+++ components/UnderConstruction.tsx
@@ -4,7 +4,14 @@ export function UnderConstruction() {
   return (
     <div className="flex min-h-screen flex-col px-4 py-12">
       <div className="mb-8 flex flex-col items-center">
-        <svg width="80" height="80" fill="none" viewBox="0 0 80 80" className="mb-4">
+        <svg
+          width="80"
+          height="80"
+          fill="none"
+          viewBox="0 0 80 80"
+          className="mb-4"
+          aria-hidden="true"
+        >
           <rect width="80" height="80" rx="16" fill="#E5E7EB" />
           <path
             d="M20 60L60 20M20 20l40 40"
